requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.3.3
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
setuptools==69.0.2
sqlparse==0.4.4
uritemplate==4.1.1
